Doing astrometry
We used Raul's astrometry make file (build-astrometry.mk) to create
our own make file, and we did some changes. We added this line:
--scale-low 2.7 --scale-high 2.9 --scale-units arcsecperpix  \
instead of $(SOLVE_FIELD_PARAMS) and changed the $(visual-good-astr-bnames)
with the basename we introduced in bias make file.
That is: $(basenames)

Also some changes we appied to the final line as bellow:
$(mtexdir)/build-astrometry.tex: $(sip-headers) $(ca-astr-ims) | $(mtexdir)

Now we can do astrometry of the images.

At the next step we installed swarp to be able to coadd the images.
To do this we first add the swarp to TARGETS.mk (at: /home/haltafi
/projects/inola/reproduce/software/config/installation) then run
./configure -e